Delete IOAkkaApp.  Delete catsakka akka operations.  Rename catsakka to catsutils.  Remove akka dependencies. (#2955)
