+++ cats-utils/src/main/scala/co/topl/catsakka/FOps.scala
@@ -1,4 +1,4 @@
-package co.topl.catsakka
+package co.topl.catsutils
 
 import cats.Monad
 import cats.effect.Clock
