+++ common-application/src/main/scala/co/topl/common/application/IOBaseApp.scala
@@ -2,7 +2,6 @@ package co.topl.common.application
 
 import cats.data.Chain
 import cats.effect._
-import cats.effect.unsafe.IORuntime
 import cats.implicits._
 import cats.kernel.Monoid
 import com.typesafe.config.{Config, ConfigFactory}
@@ -11,9 +10,8 @@ import pureconfig.{ConfigObjectSource, ConfigSource}
 import java.nio.file.Paths
 
 /**
- * Assists with constructing applications which use both cats-effect and akka.  Constructs an
- * ActorSystem that will be installed as the cats-effect runtime.  Then runs the defined
- * IO program before cleaning up.
+ * Assists with constructing applications which use cats-effect  Initializes the runtime and configurations.
+ * Then runs the defined IO program before cleaning up.
  * @param createArgs a function which turns stringified command-line args into a structured type
  * @param createConfig a function which creates a HOCON config using the parsed command-line args
  * @param parseConfig a function which create application config for given command line arguments and config
@@ -33,7 +31,6 @@ abstract class IOBaseApp[CmdArgs, AppConfig](
   protected[application] var _config: Config = _
   protected[application] var _appConfig: AppConfig = _
   protected[application] var _ioApp: IOApp = _
-  protected[application] var _ioRuntime: IORuntime = _
 
   implicit def args: CmdArgs = _args
   implicit def appConfig: AppConfig = _appConfig
@@ -41,18 +38,12 @@ abstract class IOBaseApp[CmdArgs, AppConfig](
 
   def run: IO[Unit]
 
-  protected def initRuntime(): Unit =
-    _ioRuntime = cats.effect.unsafe.IORuntime.global
-
   def initialize(args: Array[String]): Unit = {
     _args = createArgs(args.toList)
     _config = createConfig(_args)
     _appConfig = parseConfig(_args, _config)
     preInitFunction(_appConfig)
-    initRuntime()
     _ioApp = new IOApp {
-      override protected val runtime: IORuntime = _ioRuntime
-
       def run(args: List[String]): IO[ExitCode] =
         IOBaseApp.this.run.as(ExitCode.Success)
     }
