+++ minting/src/main/scala/co/topl/minting/interpreters/BlockProducer.scala
@@ -6,7 +6,7 @@ import cats.effect._
 import cats.implicits._
 import co.topl.algebras.ClockAlgebra
 import co.topl.brambl.syntax._
-import co.topl.catsakka._
+import co.topl.catsutils._
 import co.topl.consensus.models.BlockId
 import co.topl.consensus.models.StakingAddress
 import co.topl.minting.algebras.{BlockPackerAlgebra, BlockProducerAlgebra, StakingAlgebra}
