+++ node/src/main/scala/co/topl/node/NodeApp.scala
@@ -1,7 +1,5 @@
 package co.topl.node
 
-import akka.actor.typed.ActorSystem
-import akka.actor.typed.scaladsl.Behaviors
 import cats.Applicative
 import cats.effect.implicits._
 import cats.effect.std.Random
@@ -11,7 +9,7 @@ import cats.implicits._
 import co.topl.algebras._
 import co.topl.blockchain._
 import co.topl.codecs.bytes.tetra.instances._
-import co.topl.common.application.{IOAkkaApp, IOBaseApp}
+import co.topl.common.application.IOBaseApp
 import co.topl.consensus.algebras._
 import co.topl.consensus.models.VrfConfig
 import co.topl.consensus.interpreters._
@@ -43,17 +41,16 @@ import java.util.UUID
 object NodeApp extends AbstractNodeApp
 
 abstract class AbstractNodeApp
-    extends IOAkkaApp[Args, ApplicationConfig, Nothing](
+    extends IOBaseApp[Args, ApplicationConfig](
       createArgs = args => Args.parserArgs.constructOrThrow(args),
       createConfig = IOBaseApp.createTypesafeConfig,
       parseConfig = (args, conf) => ApplicationConfig.unsafe(args, conf),
-      createSystem = (_, _, conf) => ActorSystem[Nothing](Behaviors.empty, "Bifrost", conf),
       preInitFunction = config => if (config.kamon.enable) Kamon.init()
     ) {
   def run: IO[Unit] = new ConfiguredNodeApp(args, appConfig).run
 }
 
-class ConfiguredNodeApp(args: Args, appConfig: ApplicationConfig)(implicit system: ActorSystem[_]) {
+class ConfiguredNodeApp(args: Args, appConfig: ApplicationConfig) {
 
   type F[A] = IO[A]
