+++ project/Dependencies.scala
@@ -2,7 +2,6 @@ import sbt._
 
 object Dependencies {
 
-  val akkaVersion = "2.6.20"
   val circeVersion = "0.14.5"
   val kamonVersion = "2.5.12"
   val simulacrumVersion = "1.0.1"
@@ -52,9 +51,6 @@ object Dependencies {
   val mUnitTest: Seq[ModuleID] = mUnitTestBase.map(_ % Test)
 
   val dockerClient = "com.spotify" % "docker-client" % "8.16.0"
-
-  def akka(name: String): ModuleID =
-    "com.typesafe.akka" %% s"akka-$name" % akkaVersion
 
   val circe: Seq[ModuleID] = Seq(
     "io.circe" %% "circe-core"    % circeVersion,
@@ -136,7 +132,6 @@ object Dependencies {
   val node: Seq[ModuleID] =
     Seq(
       catsSlf4j,
-      akka("actor-typed"),
       fs2Core,
       fs2IO
     ) ++
@@ -179,7 +174,6 @@ object Dependencies {
     cats ++ catsEffect ++ mainargs ++ logging ++ monocle ++
     simulacrum ++ Seq(
       catsSlf4j,
-      akka("actor-typed"),
       pureConfig,
       circeYaml
     )
@@ -195,10 +189,8 @@ object Dependencies {
   lazy val eventTree: Seq[ModuleID] =
     Dependencies.mUnitTest ++ Dependencies.catsEffect
 
-  lazy val catsAkka: Seq[ModuleID] =
-    cats ++ catsEffect ++ logging ++
-    Seq(akka("actor"), akka("actor-typed"), akka("stream")) ++
-    Seq(fs2Core, fs2IO, fs2ReactiveStreams)
+  lazy val catsUtils: Seq[ModuleID] =
+    cats ++ catsEffect ++ logging ++ Seq(fs2Core, fs2IO, fs2ReactiveStreams)
 
   lazy val models: Seq[ModuleID] =
     cats ++ simulacrum ++ newType ++ scodec ++ protobufSpecs ++
@@ -206,16 +198,13 @@ object Dependencies {
     Seq(quivr4s, quivr4s.classifier("tests") % Test)
 
   lazy val consensus: Seq[ModuleID] =
-    Dependencies.mUnitTest ++ externalCrypto ++ Seq(akka("actor-typed")) ++ catsEffect ++ logging ++ scalacache
+    Dependencies.mUnitTest ++ externalCrypto ++ catsEffect ++ logging ++ scalacache
 
   lazy val minting: Seq[ModuleID] =
     Dependencies.mUnitTest ++ Dependencies.test ++ Dependencies.catsEffect
 
   lazy val networking: Seq[ModuleID] =
-    Dependencies.test ++ Dependencies.catsEffect ++ Seq(
-      Dependencies.akka("stream"),
-      Dependencies.akka("stream-testkit") % Test
-    )
+    Dependencies.test ++ Dependencies.catsEffect
 
   lazy val transactionGenerator: Seq[ModuleID] =
     Dependencies.mUnitTest ++ Dependencies.catsEffect ++ Seq(Dependencies.fs2Core)
@@ -225,10 +214,7 @@ object Dependencies {
     Seq(Dependencies.bramblScSdk, Dependencies.bramblScSdk.classifier("tests") % Test)
 
   lazy val blockchain: Seq[ModuleID] =
-    Dependencies.mUnitTest ++ Dependencies.catsEffect ++ logging ++ Seq(
-      akka("stream"),
-      akka("stream-testkit") % Test
-    ) ++ Seq(fs2Core)
+    Dependencies.mUnitTest ++ Dependencies.catsEffect ++ logging ++ Seq(fs2Core)
 
   lazy val commonInterpreters: Seq[sbt.ModuleID] =
     mUnitTest ++
@@ -244,8 +230,7 @@ object Dependencies {
     test ++
     simulacrum ++
     scodec ++
-    cats ++
-    Seq(akka("actor"))
+    cats
 
   lazy val toplGrpc: Seq[ModuleID] =
     cats ++
